+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changed:
     output file group is targeted but for pages that aren't in that group, #471
   * `ocrd_cli_wrap_processor`: Take `page_id` into account when doing `WorkspaceValidator.check_file_grp`
   * `run_cli` accepts an `overwrite` parameter to pass on to processor calls, #471
+  * `Task.validate`: set implicit input/output groups from ocrd-tool.json, #471
 
 Added: